Added interceptor that wakes sources
This interceptor wakes the sources each time there is a call to the endpoint. It also allows the intercepted handlers to wait for the sources to be ready if required by calling the `WaitForSources` function on the context. This means that no time is wasted waiting for the sources if it isn't absolutely required