Fix logic error when using bypassed auth
In the previous logic if there were any paths that bypassed authentication the check JWT middleware would not be executed which means the token would not be validated and the token information would not be extracted into the context. This means that any services that were checking things from context like whether scopes existed would find them empty. This is what caused the authentication issues that we saw in production.

We are now correctly checking if the path is bypassed for each request (rather than just checking if this is set at all), and then if it is not bypassed we are checking the JWT token, otherwise we are just passing through the request.